Added Shop and new Background
Enemies can now go through each other
There are three types of chests in the shop
There is a script to buy something in the shop
There is an array which detects all chests inside the child collider
The chest that is closest to the player will be bought.
Currently no inventory but the player does lose coins when buying.