new Animation and background
the player no longer enters the house as soon as he touches it, he has to press "Q". Pressing q near multiple doors will create an array of houses and iterate through them to find the closest one. A similar approach was done to the chests in the shop scene